[Ref] Rename DEFAULT_PAGE_SIZE constant to DEFAULT_TREE_SELECT_FETCH_SIZE, move getInitialPageSize function to UISettingsUtil and make it fall back to middle value of available page sizes
