+++ src/component/misc/table/Pagination.tsx
@@ -34,9 +34,7 @@ interface PaginationProps {
   allowSizeChange?: boolean;
 }
 
-const PAGE_SIZES = [10, 20, 30, 50, Constants.DEFAULT_PAGE_SIZE].sort(
-  (a, b) => a - b // Sort by number value
-);
+export const PAGE_SIZES = [10, 20, 30, 50];
 
 export const Pagination: React.FC<PaginationProps> = ({
   table,
